+++ givenergy_modbus/pdu/codec.py
